css/css-page/page-rule-declarations-001.html

<!doctype html>
<html>
<head>
    <title>CSS Paged Media: parsing @page declarations inside @media</title>
    <link rel="author" title="Felipe Erias Morandeira" href="mailto:felipeerias@gmail.com"/>
    <link rel="help" href="https://www.w3.org/TR/css-page-3"/>
    <meta name="assert" content="Test that @page declarations inside @media are parsed correctly.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/parsing-testcommon.js"></script>

    <style type="text/css">

        @media print {
            @page {
                margin: 3cm;
            }
            @page :first {
                margin-top: 6cm;
            }
            @page :left {
                color: red;
            }
            @page :right {
                color: blue;
            }
        }

    </style>
    <script type="text/javascript">

        test(function(){
            let expectedForSelector = {
                "" : "margin: 3cm;",
                ":first" : "margin-top: 6cm;",
                ":left" : "color: red;",
                ":right" : "color: blue;"
            };
            let ss = document.styleSheets;
            for (let i = 0; i < ss.length; i++) {
                let rules = ss[i].cssRules;
                for (let rule of rules) {
                    if (rule.type == CSSRule.MEDIA_RULE && rule.conditionText == 'print') {
                        for (let mediaRule of rule.cssRules) {
                            if (mediaRule.type == CSSRule.PAGE_RULE) {
                                let expected = expectedForSelector[mediaRule.selectorText];
                                assert_equals(mediaRule.style.cssText, expected, "unexpected @page contents");
                                delete expectedForSelector[mediaRule.selectorText];
                            }
                        }
                    }
                }
            }
            assert_equals(Object.keys(expectedForSelector).length, 0, "missing @page selectors in @media");
        }, "test @page inside a @media print rule");

    </script>
</head>
<body>
</body>
</html>
